Move common IPC code to Hosting project
