--- src/Microsoft.MobileBlazorBindings.Hosting/IPCBase.cs
@@ -1,10 +1,0 @@
-﻿// Copyright (c) Microsoft Corporation.
-// Licensed under the MIT license.
-
-namespace Microsoft.MobileBlazorBindings.Hosting
-{
-    public abstract class IPCBase
-    {
-        public abstract void Send(string eventName, params object[] args);
-    }
-}
